create a new iou utils function
